requirements.txt
fastapi
uvicorn
praw
supabase
python-dotenv
spacy
spacytextblob